+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -150,6 +150,12 @@ New checks
   Finds initializer lists for aggregate types that could be
   written as designated initializers instead.
 
+- New :doc:`modernize-use-uniform-initializer
+  <clang-tidy/checks/modernize/use-uniform-initializer>` check.
+
+  Finds usage of C-Style initialization that can be rewritten with
+  C++-11 uniform initializers.
+
 - New :doc:`readability-enum-initial-value
   <clang-tidy/checks/readability/enum-initial-value>` check.